font-patcher: Use correct source font metrics
[why]
With a source font where Win Ascent/Descent differs from Typo
Ascend/Descent newly added symbols that are intended to be centered
_within the visual space_ can end up too far up or down.

The happens for example when patching CascadiaCode. Added glyphs like
the Ubuntu logo (unicode 0xF31B) is not centered between the square
brackets or on a line with the less then and other centered glyphs.

[how]
The calculation takes the Win Ascent/Descent to calculate the visual
hight. That information is a mix of hight and line spacing and can be
misleading.

Therefore, if use_typo_metrics is set in a font, we obey that flag
and use the typo metrics values instead.

[note]
Some websites with further information follow.

https://github.com/googlefonts/gf-docs/tree/main/VerticalMetrics
  > Hhea metrics are used in Mac OS X, whilst Microsoft uses
  > Typo when Use_Typo_Metrics is enabled

https://docs.microsoft.com/en-us/typography/opentype/otspec160/os2#fsselection
https://developer.apple.com/fonts/TrueType-Reference-Manual/RM06/Chap6hhea.html

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>