doc: Correct example patch call through fontforge [skip ci]
[why]
The readme states that one should call `fontforge` with `./fortforge`.
That means that no PATH is used, and will usually fail, because the user
is in our repo and not in the directory where the fontforge binary
resides in. And even if ... then the `font-patcher` script would be
somewhere else. That makes no sense.

[how]
Drop the `./` prefix of the `fontforge` call.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>