Clean up icon doubles
Some icons are now existing twice.

* Their crystal icon looks wrong (rotationwise), compare with https://github.com/crystal-lang/crystal
* Their C logo misses the standard's hex shape
* Their C++ logo misses the standard's hex shape
* Their Elixir icons has a dent in the reflextion?
* Their Elm icon is colored
* Their Git Folder is not a folder
* Their Kotlin icon is colored
* Their Puppet icon is more correct
* Their Pursescript icon is identical
* Their Windows icon is more modern

That means:

* Do not include their crystal, elixir, elm, kotlin, purescript
* Do not include our puppet, windows
* Keeping both c, c++, git-folder

[note]
Keeping their old files around with changed suffix to find this commit
in the future on updates.

Reuse old icons that were labeled as 'seti' but have been vanished as
new custom types with appropriate file naming.

And remove 'red' versions of icons. We have a black and white font...

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>