doc: Fix RFN entries in readme [skip ci]
[why]
Some fonts in the readme source font tables are marked as 'no RFN' while
they infact have a RFN (and we act on that, so that is just a
documentation issue).

[how]
Also add updated fonts.json, that contains RFN information and correct
unpatchedFont names, compare gh-pages commit
  4f78d656  download: Fix RFN detection

The updated fonts.json is also needed for an upcoming Casks creation
fix.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>